Client/requirement.txt
textual
textual-dev
python-socketio[asyncio_client]